requirements.txt
mcp>=1.9.0
starlette>=0.27.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
atlassian-python-api>=3.41.0